convert from using tokens to using api keys (#39)
* convert from using tokens to using api keys

* convert frontend to only use api keys instead of refresh tokens

* call the logout endpoint

* Add back google authentication

* fix lint

* add otp wrapper

* add config fields

+++ tests/api/test_users.py
@@ -14,68 +14,54 @@ def test_user_auth_functions(app_client: TestClient, mock_send_email: MockType) 
 
     test_email = "test@example.com"
     login_url = "/"
-    bad_actor_email = "badactor@gmail.com"
 
-    # Creates a bad actor user for testing admin actions later.
-    otp = OneTimePassPayload(email=bad_actor_email)
-    response = app_client.post("/api/users/otp", json={"payload": otp.encode()})
-    assert response.status_code == 200, response.json()
-
-    # Sends an email to the user with their one-time pass.
-    response = app_client.post(
-        "/api/users/login",
-        json={
-            "email": test_email,
-            "login_url": login_url,
-        },
-    )
+    # Sends the one-time password to the test email.
+    response = app_client.post("/api/users/login", json={"email": test_email, "login_url": login_url})
     assert response.status_code == 200, response.json()
     assert mock_send_email.call_count == 1
 
-    # Uses the one-time pass to set client cookies.
+    # Uses the one-time pass to get an API key. We need to make a new OTP
+    # manually because we can't send emails in unit tests.
     otp = OneTimePassPayload(email=test_email)
     response = app_client.post("/api/users/otp", json={"payload": otp.encode()})
     assert response.status_code == 200, response.json()
+    response_data = response.json()
+    api_key = response_data["api_key"]
 
-    # Checks that we get a 401 without a session token.
+    # Checks that without the API key we get a 401 response.
     response = app_client.get("/api/users/me")
     assert response.status_code == 401, response.json()
     assert response.json()["detail"] == "Not authenticated"
 
-    # Get a session token.
-    response = app_client.post("/api/users/refresh")
-    assert response.status_code == 200, response.json()
-    assert response.json()["token_type"] == "Bearer"
-
-    # Gets the user's profile using the token.
-    response = app_client.get("/api/users/me")
+    # Checks that with the API key we get a 200 response.
+    response = app_client.get("/api/users/me", headers={"Authorization": f"Bearer {api_key}"})
     assert response.status_code == 200, response.json()
     assert response.json()["email"] == test_email
 
-    # Log the user out.
+    # Checks that we can't log the user out without the API key.
     response = app_client.delete("/api/users/logout")
+    assert response.status_code == 401, response.json()
+
+    # Log the user out, which deletes the API key.
+    response = app_client.delete("/api/users/logout", headers={"Authorization": f"Bearer {api_key}"})
     assert response.status_code == 200, response.json()
     assert response.json() is True
 
-    # Check that the user cookie has been cleared.
-    response = app_client.get("/api/users/me")
-    assert response.status_code == 401, response.json()
-    assert response.json()["detail"] == "Not authenticated"
+    # Checks that we can no longer use that API key to get the user's info.
+    response = app_client.get("/api/users/me", headers={"Authorization": f"Bearer {api_key}"})
+    assert response.status_code == 404, response.json()
+    assert response.json()["detail"] == "User not found"
 
-    # Log the user back in.
+    # Log the user back in, getting new API key.
     response = app_client.post("/api/users/otp", json={"payload": otp.encode()})
     assert response.status_code == 200, response.json()
 
-    # Gets another session token.
-    response = app_client.post("/api/users/refresh")
-    assert response.status_code == 200, response.json()
-
-    # Delete the user.
-    response = app_client.delete("/api/users/me")
+    # Delete the user using the new API key.
+    response = app_client.delete("/api/users/me", headers={"Authorization": f"Bearer {api_key}"})
     assert response.status_code == 200, response.json()
     assert response.json() is True
 
-    # Make sure the user is gone.
-    response = app_client.get("/api/users/me")
-    assert response.status_code == 400, response.json()
+    # Tries deleting the user again, which should fail.
+    response = app_client.delete("/api/users/me", headers={"Authorization": f"Bearer {api_key}"})
+    assert response.status_code == 404, response.json()
     assert response.json()["detail"] == "User not found"
